+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Benchmarks/Models/BenchmarkDefaults.cs
@@ -1,0 +1,12 @@
+namespace Azure.Sdk.Tools.Cli.Benchmarks.Models;
+
+/// <summary>
+/// Default values for benchmark execution.
+/// </summary>
+public static class BenchmarkDefaults
+{
+    /// <summary>
+    /// The default model to use for benchmark execution.
+    /// </summary>
+    public const string DefaultModel = "claude-opus-4.5";
+}
