+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Benchmarks/Models/RepoConfig.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-namespace Azure.Sdk.Tools.Cli.Benchmarks.Infrastructure;
+namespace Azure.Sdk.Tools.Cli.Benchmarks.Models;
 
 /// <summary>
 /// Configuration for a GitHub repository used in benchmarks.
