+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Benchmarks/Scenarios/BenchmarkScenario.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using Azure.Sdk.Tools.Cli.Benchmarks.Infrastructure;
+using Azure.Sdk.Tools.Cli.Benchmarks.Models;
 using Azure.Sdk.Tools.Cli.Benchmarks.Validation;
 
 namespace Azure.Sdk.Tools.Cli.Benchmarks.Scenarios;
