+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Benchmarks/Validation/LlmJudge.cs
@@ -1,25 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Azure.Sdk.Tools.Cli.Benchmarks.Models;
 using GitHub.Copilot.SDK;
 
 namespace Azure.Sdk.Tools.Cli.Benchmarks.Validation;
-
-/// <summary>
-/// Result of an LLM judgment.
-/// </summary>
-public class JudgmentResult
-{
-    /// <summary>
-    /// Gets whether the judgment passed.
-    /// </summary>
-    public required bool Passed { get; init; }
-
-    /// <summary>
-    /// Gets the reasoning provided by the LLM.
-    /// </summary>
-    public required string Reasoning { get; init; }
-}
 
 /// <summary>
 /// Helper for making LLM judgment calls using the GitHub Copilot SDK.
